ChatAgent/pyproject.toml
[project]
name = "advanced-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firecrawl-py>=2.7.1",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.19",
    "langgraph>=0.4.8",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]